+++ extensions/Worker.Extensions.Tables/src/Constants.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Microsoft.Azure.Functions.Worker.Extensions.Tables
+{
+    internal class Constants
+    {
+        internal const string Storage = "Storage";
+        internal const string TablesExtensionName = "AzureStorageTables";
+        internal const string TableName = "TableName";
+        internal const string PartitionKey = "PartitionKey";
+        internal const string RowKey = "RowKey";
+        internal const string Connection = "Connection";
+        internal const string Take = "Take";
+        internal const string Filter = "Filter";
+        // Media content types
+        internal const string JsonContentType = "application/json";
+    }
+}